gRPCClient.jl

gRPCClient.jl aims to be a production grade gRPC client emphasizing performance and reliability.

Features

Unary+Streaming RPC

HTTP/2 connection multiplexing

Synchronous and asynchronous interfaces

Thread safe

SSL/TLS

The client is missing a few features which will be added over time if there is sufficient interest:

OAuth2

Compression

Getting Started

Test gRPC Server

All examples in the documentation are run against a test server written in Go. You can run it by doing the following:

cd test/go

# Build
go build -o grpc_test_server

# Run
./grpc_test_server

Code Generation

Note: support for this is currently being upstreamed into ProtoBuf.jl. Until then, make sure you add the feature branch with gRPC code generation support:

pkg> add https://github.com/csvance/ProtoBuf.jl#external-service-support

gRPCClient.jl integrates with ProtoBuf.jl to automatically generate Julia client stubs for calling gRPC.

using ProtoBuf
using gRPCClient

# Register our service codegen with ProtoBuf.jl
grpc_register_service_codegen()

# Creates Julia bindings for the messages and RPC defined in test.proto
protojl("test/proto/test.proto", ".", "test/gen")

Example Usage

See here for examples covering all provided interfaces for both unary and streaming gRPC calls.

API

Package Initialization / Shutdown

gRPCClient.grpc_init — Method

grpc_init([grpc_curl::gRPCCURL])

Initializes the gRPCCURL object. This should be called once before making gRPC calls. There is no harm in calling this more than once (ie by different packages/dependencies). Typical usage looks like this:

grpc_init()

client = TestService_TestRPC_Client("172.238.177.88", 8001)

# Make some gRPC calls 

# Shut down the global gRPC handle
grpc_shutdown()

Unless specifying a gRPCCURL the global one provided by grpc_global_handle() is used. Each gRPCCURL state has its own connection pool and request semaphore, so sometimes you may want to manage your own like shown below:

grpc_myapp = gRPCCURL()
grpc_init(grpc_myapp)

client = TestService_TestRPC_Client("172.238.177.88", 8001; grpc=grpc_myapp)

# Make some gRPC calls 

# Only shuts down your gRPC handle
grpc_shutdown(grpc_myapp)

gRPCClient.grpc_shutdown — Method

grpc_shutdown([grpc_curl::gRPCCURL])

Shuts down the gRPCCURL. This neatly cleans up all active connections and requests. Useful for calling during development with Revise. Unless specifying the gRPCCURL, the global one provided by grpc_global_handle() is shutdown.

gRPCClient.grpc_global_handle — Method

grpc_global_handle()

Returns the global gRPCCURL state which contains a libCURL multi handle. By default all gRPC clients use this multi in order to ensure that HTTP/2 multiplexing happens where possible.

RPC

Unary

gRPCClient.grpc_async_request — Method

grpc_async_request(client::gRPCServiceClient{TRequest,false,TResponse,false}, request::TRequest) where {TRequest<:Any,TResponse<:Any}

Initiate an asynchronous gRPC request: send the request to the server and then immediately return a gRPCRequest object without waiting for the response. In order to wait on / retrieve the result once its ready, call grpc_async_await. This is ideal when you need to send many requests in parallel and waiting on each response before sending the next request would things down.

using gRPCClient

grpc_init()

# Include the generated bindings
include("test/gen/test/test_pb.jl")

# Create a client bound to a specific RPC
client = TestService_TestRPC_Client("localhost", 8001)

# Make a syncronous request and get back a TestResponse
response = grpc_sync_request(client, TestRequest(1, zeros(UInt64, 1)))
@info response

# Make some async requests and await their TestResponse
requests = Vector{gRPCRequest}()
for i in 1:10
    push!(
        requests, 
        grpc_async_request(client, TestRequest(1, zeros(UInt64, 1)))
    )
end

for request in requests
    response = grpc_async_await(client, request)
    @info response
end

gRPCClient.grpc_async_request — Method

grpc_async_request(client::gRPCServiceClient{TRequest,false,TResponse,false}, request::TRequest, channel::Channel{gRPCAsyncChannelResponse{TResponse}}, index::Int64) where {TRequest<:Any,TResponse<:Any}

Initiate an asynchronous gRPC request: send the request to the server and then immediately return. When the request is complete a background task will put the response in the provided channel. This has the advantage over the request / await patern in that you can handle responses immediately after they are recieved in any order.

using gRPCClient

grpc_init()
include("test/gen/test/test_pb.jl")

# Connect to the test server
client = TestService_TestRPC_Client("localhost", 8001)

N = 10

channel = Channel{gRPCAsyncChannelResponse{TestResponse}}(N)

for (index, request) in enumerate([TestRequest(i, zeros(UInt64, i)) for i in 1:N])
     grpc_async_request(client, request, channel, index)
end

for i in 1:N
    cr = take!(channel)
    # Check if an exception was thrown, if so throw it here
    !isnothing(cr.ex) && throw(cr.ex)

    # If this does not hold true, then the requests and responses have gotten mixed up.
    @assert length(cr.response.data) == cr.index
end

gRPCClient.grpc_async_await — Method

grpc_async_await(client::gRPCServiceClient{TRequest,false,TResponse,false}, request::gRPCRequest) where {TRequest<:Any,TResponse<:Any}

Wait for the request to complete and return the response when it is ready. Throws any exceptions that were encountered during handling of the request.

gRPCClient.grpc_sync_request — Method

grpc_sync_request(client::gRPCServiceClient{TRequest,false,TResponse,false}, request::TRequest) where {TRequest<:Any,TResponse<:Any}

Do a synchronous gRPC request: send the request and wait for the response before returning it. Under the hood this just calls grpc_async_request and grpc_async_await

Streaming

gRPCClient.grpc_async_request — Method

grpc_async_request(client::gRPCServiceClient{TRequest,true,TResponse,false}, request::Channel{TRequest}) where {TRequest<:Any,TResponse<:Any}

Start a requesting streaming gRPC request.

using gRPCClient

grpc_init()
include("test/gen/test/test_pb.jl")

client = TestService_TestClientStreamRPC_Client("localhost", 8001)
request_c = Channel{TestRequest}(16)
put!(request_c, TestRequest(1, zeros(UInt64, 1)))

req = grpc_async_request(client, request_c)

# Must close the request channel when done sending requests
close(request_c)

# Get the response
test_response = grpc_async_await(client, req)

gRPCClient.grpc_async_request — Method

grpc_async_request(client::gRPCServiceClient{TRequest,false,TResponse,true},request::TRequest,response::Channel{TResponse}) where {TRequest<:Any,TResponse<:Any}

Start a response streaming gRPC request.

using gRPCClient

grpc_init()
include("test/gen/test/test_pb.jl")

client = TestService_TestServerStreamRPC_Client("localhost", 8001)

response_c = Channel{TestResponse}(16)

req = grpc_async_request(
    client,
    TestRequest(1, zeros(UInt64, 1)),
    response_c,
)
test_response = take!(response_c)

# Raise any exceptions encountered during the request
grpc_async_await(req) 

gRPCClient.grpc_async_request — Method

grpc_async_request(client::gRPCServiceClient{TRequest,true,TResponse,true},request::Channel{TRequest},response::Channel{TResponse}) where {TRequest<:Any,TResponse<:Any}

Start a bidirectional gRPC request.

using gRPCClient

grpc_init()
include("test/gen/test/test_pb.jl")

client = TestService_TestBidirectionalStreamRPC_Client("localhost", 8001)

request_c = Channel{TestRequest}(16)
response_c = Channel{TestResponse}(16)

put!(request_c, TestRequest(1, zeros(UInt64, 1)))
req = grpc_async_request(client, request_c, response_c)
test_response = take!(response_c)

# Must close the request channel when done sending requests
close(request_c)
# Raise any exceptions encountered during the request
grpc_async_await(req) 

gRPCClient.grpc_async_await — Method

grpc_async_await(client::gRPCServiceClient{TRequest,true,TResponse,false},request::gRPCRequest) where {TRequest<:Any,TResponse<:Any}

Raise any exceptions encountered during the streaming request.

Exceptions

gRPCClient.gRPCServiceCallException — Type

Exception type that is thrown when something goes wrong while calling an RPC. This can either be triggered by the servers response code or by the client when something fails.

This exception type has two fields:

grpc_status::Int - See here for an indepth explanation of each status.

message::String

Benchmarking

All benchmark tests run against the Test gRPC Server in test/go. See the relevant documentation for information on how to run this.

TODO: information on how to run the benchmarks with PrettyTables.jl support

Stress Workloads

In addition to benchmarks, a number of workloads based on these are available in workloads.jl:

stress_workload_smol()

stress_workload_32_224_224_uint8()

stress_workload_streaming_request()

stress_workload_streaming_response()

stress_workload_streaming_bidirectional()

These run forever, and are useful to help identify any stability issues or resource leaks.